Rename sketch.h -> shared.h
