+++ tests/plugins/Qukeys/basic/test/testcase.cpp
@@ -16,7 +16,7 @@
 
 #include "testing/setup-googletest.h"
 
-#include "../sketch.h"
+#include "../shared.h"
 
 SETUP_GOOGLETEST();
